requirements.txt
asgiref==3.8.1
astroid==3.3.9
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
crispy-bootstrap5==2024.10
dill==0.3.9
Django==4.2.20
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-environ==0.12.0
django-filter==25.1
django-formtools==2.5.1
django-storages==1.14.5
djangorestframework==3.15.2
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
isort==6.0.1
mccabe==0.7.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pylint==3.3.5
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
